Convert runner/plugin.js to typescript
